♻️ Move helpers to components
